Add CapitalizeDecorator and TrimmerDecorator classes
