+++ main.rb
@@ -1,0 +1,9 @@
+require_relative 'person'
+require_relative 'decorators'
+
+person = Person.new(5, 'maisythedog')
+puts person.correct_name
+capitalized_person = CapitalizeDecorator.new(person)
+puts capitalized_person.correct_name
+capitalized_trimmed_person = TrimmerDecorator.new(capitalized_person)
+puts capitalized_trimmed_person.correct_name
