+++ student.rb
@@ -10,5 +10,3 @@ class Student < Person
     "¯\(ツ)/¯"
   end
 end
-
-
